refactor(responsecache): split stream cache by API direction
internal/responsecache/stream_cache.go was 1309 lines bundling SSE
helpers, chat-stream cache builder, and responses-stream cache
builder. Split into peer files in the same package:

  stream_cache.go (304)            entry helpers — cacheKeyRequestBody,
                                   writeCachedResponse, SSE parsing
                                   (parseSSEJSONEvents, parseCacheEventJSON,
                                   nextCacheEventBoundary, parseCacheDataLine),
                                   reconstructStreamingResponse dispatcher,
                                   stream-options helpers, JSON helpers
  stream_cache_chat.go (395)       chatStreamCacheBuilder + state types,
                                   renderCachedChatStream,
                                   buildChatToolCalls, renderChatToolCalls,
                                   chatUsageMap, chatReasoningContent
  stream_cache_responses.go (627)  responsesStreamCacheBuilder + state types,
                                   renderCachedResponsesStream,
                                   responsesAddedItem, responsesTerminalEventName,
                                   appendResponsesItemDeltaEvents,
                                   responsesContentDeltaEvent

Pure relocation; tests pass.
